+++ src/main/java/module-info.java
@@ -1,4 +1,8 @@
 module io.github.parseworks {
     exports io.github.parseworks;
     exports io.github.parseworks.impl;
+    exports io.github.parseworks.plugin;
+    exports io.github.parseworks.plugin.examples;
+
+    uses io.github.parseworks.plugin.ParserPlugin;
 }